Add template func for rendering netbox config
- Refactor netbox code to pkg
